% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bedpe2gene.R
\name{annotate_bedpe2gene}
\alias{annotate_bedpe2gene}
\title{annotate_bedpe2gene}
\usage{
annotate_bedpe2gene(
  bedpe,
  prom_bed,
  gene_side = NULL,
  bedpe_chr_a = "chr_a",
  bedpe_start_a = "start_a",
  bedpe_end_a = "end_a",
  bedpe_chr_b = "chr_b",
  bedpe_start_b = "start_b",
  bedpe_end_b = "end_b",
  prom_chr = "pro_chr",
  prom_start = "pro_start",
  prom_end = "pro_end",
  prom_anno = "pro_anno"
)
}
\arguments{
\item{bedpe}{bedpe data.frame}

\item{prom_bed}{prom_bed data.frame}

\item{gene_side}{NULL or bedpe column name match (eg."bait");
if NULL, search genes on both ends of bedpe.}

\item{bedpe_chr_a}{character}

\item{bedpe_start_a}{character}

\item{bedpe_end_a}{character}

\item{bedpe_chr_b}{character}

\item{bedpe_start_b}{character}

\item{bedpe_end_b}{character}

\item{prom_chr}{character}

\item{prom_start}{character}

\item{prom_end}{character}

\item{prom_anno}{character}
}
\value{
data.frame with a and b end annotated to pro_anno
}
\description{
annotate bedpe loop with one end as gene promoter
}
\examples{
# annotate bedpe on both sides

\donttest{data(bedpe)
data(prom_bed)
annotate_bedpe2gene(bedpe, prom_bed)
}

# annotate ibed file on one side

\donttest{data(ibed)
data(prom_bed)
annotate_bedpe2gene(
ibed \%>\% mutate_at(vars(ends_with("_start")), function(x){x=x-1}),
prom_bed, gene_side="bait",
bedpe_chr_a="bait_chr", bedpe_start_a="bait_start", bedpe_end_a="bait_end",
bedpe_chr_b="otherEnd_chr", bedpe_start_b="otherEnd_start", bedpe_end_b="otherEnd_end"
)
}


}
