% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baitmap-data.R
\docType{data}
\name{baitmap}
\alias{baitmap}
\title{baitmap data}
\format{
a data.frame with columns
including bait_chr, bait_start, bait_end, bait_name, oe_chr, oe_start
oe_end, N, score
}
\usage{
data(baitmap)
}
\description{
baitmap data is the format direct input for chicago
}
\examples{
data(baitmap)
\donttest{
head(baitmap)}
}
\author{
Chun Su
}
\keyword{datasets}
