% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedpe-data.R
\docType{data}
\name{bedpe}
\alias{bedpe}
\title{bedpe data}
\format{
a data.frame with columns includin
chr_a, start_a, end_a, chr_b, start_b, end_b, val1, val2
}
\usage{
data(bedpe)
}
\description{
bedpe is a interaction file format
}
\examples{
data(bedpe)
\donttest{
head(bedpe)}
}
\author{
Chun Su
}
\keyword{datasets}
