% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibed-data.R
\docType{data}
\name{ibed}
\alias{ibed}
\title{ibed data}
\format{
a data.frame with columns
including chr, start, end, id, anno
}
\usage{
data(ibed)
}
\description{
ibed data is the format directly from chicago
}
\examples{
data(ibed)
\donttest{
head(ibed)}
}
\author{
Chun Su
}
\keyword{datasets}
