% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prom_bed-data.R
\docType{data}
\name{prom_bed}
\alias{prom_bed}
\title{Example promoter bed data}
\format{
a data.frame with column names
pro_chr, pro_start, pro_end, pro_anno, strand and gene_id
}
\usage{
data(prom_bed)
}
\description{
prom_bed example data list the coordinates of non-pseudogene promoter region
(-1500 ~ 500bp of TSS in gencodeV19 annotation), and corresponding annotation (gene_name+transcript_id)
and gene_name
}
\examples{
data(prom_bed)
\donttest{
head(prom_bed)}
}
\author{
Chun Su
}
\keyword{datasets}
