% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ibed_with_int_id.R
\name{read_ibed_with_int_id}
\alias{read_ibed_with_int_id}
\title{Read ibed file to data.frame and add interaction id as new column}
\usage{
read_ibed_with_int_id(ibed, parse_b2b = NULL, max_score = F)
}
\arguments{
\item{ibed}{ibed file from Chicago or data.frame with ibed header
ibed header requirement: bait_chr, bait_start, bait_end, bait_name, otherEnd_chr, otherEnd_start,
otherEnd_end, otherEnd_name}

\item{parse_b2b}{convert all bait-to-bait interactions to bi-directional ("double all"), or
or only keep bi-directional bait-to-bait interactions ("only bi-direction"). default = NULL}

\item{max_score}{whether convert bait-to-bait interactions to max score of both directions.
default = F. In case of parse_b2b="double all" and max_score=T, NA will fill score and N_reads in bait-to-bait
interactions in which only one direction is significant.}
}
\value{
data.frame with extra column called "int_id"
}
\description{
A ibed file contains duplicates interaction because bait-to-bait interaction may be
doubled for both baits. Some bait-to-bait show significant at only one direction.
This function allows to detect bi-directional or uni-directional significant interactions,
and assign bi-directional with same interaction id.
}
\examples{

\donttest{
read_ibed_with_int_id(
system.file("extdata", "chicagoResults.ibed", package = "parseIbed"),
parse_b2b="only bi-direction"
)
}

data(example)
\donttest{read_ibed_with_int_id(ibed, parse_b2b="double all")}

}
