% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_washU_with_int_id.R
\name{read_washU_with_int_id}
\alias{read_washU_with_int_id}
\title{Read chicago interaction in washU format}
\usage{
read_washU_with_int_id(file)
}
\arguments{
\item{file}{chicago output in washU format}
}
\description{
Read chicago interaction in washU format
}
