% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_from_ibed.R
\name{summarise_from_ibed}
\alias{summarise_from_ibed}
\title{Summarize ibed}
\usage{
summarise_from_ibed(ibed, baitmap)
}
\arguments{
\item{ibed}{ibed file from Chicago or data.frame with ibed header
ibed header requirement: bait_chr, bait_start, bait_end, bait_name, otherEnd_chr, otherEnd_start,
otherEnd_end, otherEnd_name, int_id}

\item{baitmap}{bait design file corresponding to interaction being called in Chicago or data.frame
containing column "chr", "start", "end" and "name"|"anno"}
}
\value{
data.frame of summary stats
}
\description{
summary statistics of interactions in the context of bait, otherEnd and interaction
}
\examples{
data(example)
\donttest{summarise_from_ibed(ibed, baitmap)}


}
